+++ ui/projects/portofino/src/lib/content.component.ts
@@ -11,13 +11,14 @@ import {
 import {ActivatedRoute, Router} from "@angular/router";
 import {HttpClient} from "@angular/common/http";
 import {MainPageDirective} from "./content.directive";
-import {Subscription} from "rxjs";
+import {of, Subscription} from "rxjs";
 import {PortofinoService} from "./portofino.service";
 import {AuthenticationService} from "./security/authentication.service";
-import {Page, PageService} from "./page";
+import {Page, PageChild, PageService} from "./page";
 import {PageFactoryComponent} from "./page.factory";
 import {NotificationService} from "./notifications/notification.service";
 import {TranslateService} from "@ngx-translate/core";
+import {catchError, map, mergeMap} from "rxjs/operators";
 
 @Component({
   selector: 'portofino-content',
@@ -113,16 +114,47 @@ export class ContentComponent implements AfterViewInit, OnInit, OnDestroy {
     viewContainerRef.insert(componentRef.hostView);
     this.pageService.notifyPageLoaded(page);
     page.children.forEach(child => {
-      page.checkAccessibility(child);
+      this.checkAccessibility(page, child);
     });
     const parent = page.parent;
     if (parent) {
       parent.children.forEach(child => {
         if (`${parent.path}/${child.path}` != page.path) {
-          parent.checkAccessibility(child);
+          this.checkAccessibility(parent, child);
         }
       });
     }
   }
 
+  checkAccessibility(page: Page, child: PageChild) {
+    page.loadChildConfiguration(child).pipe(mergeMap(config => {
+      const componentType = PageFactoryComponent.components[config.type];
+      let computeSecurityCheckUrl = Page.defaultComputeSourceUrl;
+      if(componentType && componentType.computeSecurityCheckUrl) {
+        computeSecurityCheckUrl = componentType.computeSecurityCheckUrl;
+      }
+      const dummy = new DummyPage(
+        this.portofino, this.http, this.router, this.route, this.authenticationService, this.notificationService, this.translate,
+        computeSecurityCheckUrl(this.portofino.apiRoot, page, config.source));
+      dummy.parent = page;
+      dummy.configuration = config;
+      return dummy.checkAccess(false).pipe(map(() => true), catchError(() => of(false)));
+    })).subscribe(flag => child.accessible = flag);
+  }
+
 }
+
+class DummyPage extends Page {
+
+  constructor(
+    portofino: PortofinoService, http: HttpClient, router: Router, route: ActivatedRoute,
+    authenticationService: AuthenticationService, notificationService: NotificationService, translate: TranslateService,
+    public securityCheckUrl: string) {
+    super(portofino, http, router, route, authenticationService, notificationService, translate);
+  }
+
+  protected computeSecurityCheckUrl(): any {
+    return this.securityCheckUrl;
+  }
+}
+