+++ ui/projects/portofino/src/lib/pages/crud/detail/create.component.ts
@@ -1,10 +1,10 @@
 import {ChangeDetectorRef, Component, OnInit} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {PortofinoService} from "../../portofino.service";
-import {isInsertable, Property} from "../../class-accessor";
+import {PortofinoService} from "../../../portofino.service";
+import {isInsertable, Property} from "../../../class-accessor";
 import {BaseDetailComponent} from "../common.component";
-import {NotificationService} from "../../notifications/notification.service";
-import {Button} from "../../buttons";
+import {NotificationService} from "../../../notifications/notification.service";
+import {Button} from "../../../buttons";
 import {TranslateService} from "@ngx-translate/core";
 
 @Component({