+++ ui/projects/portofino/src/lib/pages/crud/search/search.component.ts
@@ -15,19 +15,19 @@ import {
   isInSummary, isPassword,
   isSearchable,
   Property, SelectionOption
-} from "../../class-accessor";
+} from "../../../class-accessor";
 import { PageEvent } from "@angular/material/paginator";
 import { Sort } from "@angular/material/sort";
 import { MatTableDataSource } from "@angular/material/table";
 import {HttpClient, HttpParams} from "@angular/common/http";
-import {PortofinoService} from "../../portofino.service";
+import {PortofinoService} from "../../../portofino.service";
 import {FormArray, FormControl, FormGroup} from "@angular/forms";
 import {SelectionModel} from "@angular/cdk/collections";
 import {SelectionProvider} from "../crud.common";
-import {AuthenticationService} from "../../security/authentication.service";
+import {AuthenticationService} from "../../../security/authentication.service";
 import {MediaObserver} from "@angular/flex-layout";
 import {Observable, of, Subject, Subscription} from "rxjs";
-import {ButtonInfo, getButtons} from "../../buttons";
+import {ButtonInfo, getButtons} from "../../../buttons";
 import {TranslateService} from "@ngx-translate/core";
 import {SearchFieldComponent} from "./search-field.component";