+++ ui/projects/portofino/src/lib/pages/many-to-many/many-to-many.component.ts
@@ -1,8 +1,8 @@
 import {Component, Input} from '@angular/core';
-import {PortofinoComponent} from "../page.factory";
+import {PortofinoComponent} from "../../page.factory";
 import {FormControl} from "@angular/forms";
-import {Operation, Page, PageConfiguration} from "../page";
-import {Button} from "../buttons";
+import {Operation, Page, PageConfiguration} from "../../page";
+import {Button} from "../../buttons";
 
 @Component({
   selector: 'portofino-many-to-many',